Updated GetReadplan() stats to report # of steps (bucketed)
This will be useful when assessing how expensive "playing" a readplan
will be in terms of there being multiple steps perhaps to the same
objects. For whole file GETs from e.g. Middleware, this is actually
expected (at least when the file/object was written randomly). For
normal file reads where the read size matches the write size, the
length of the readplan is expected to be precisely one.
