+++ inode/file.go
@@ -217,7 +217,7 @@ func (vS *volumeStruct) GetReadPlan(fileInodeNumber InodeNumber, offset *uint64,
 		return
 	}
 
-	stats.IncrementOperationsAndBucketedBytes(stats.FileReadplan, readPlanBytes)
+	stats.IncrementOperationsBuckedtedBytesAndBucketedSteps(stats.FileReadplan, readPlanBytes, uint64(len(readPlan)))
 	return
 }
