+++ stats/strings.go
@@ -94,6 +94,11 @@ var (
 	FileReadplanOps64K                = "proxyfs.inode.file.readplan.operations.size-32KB-to-64KB"
 	FileReadplanOpsOver64K            = "proxyfs.inode.file.readplan.operations.size-over-64KB"
 	FileReadplanBytes                 = "proxyfs.inode.file.readplan.bytes"
+	FileReadplanOpsSteps1             = "proxyfs.inode.file.readplan.operations.steps-1"
+	FileReadplanOpsStepsTo4           = "proxyfs.inode.file.readplan.operations.steps-2-to-4"
+	FileReadplanOpsStepsTo16          = "proxyfs.inode.file.readplan.operations.steps-5-to-16"
+	FileReadplanOpsStepsTo64          = "proxyfs.inode.file.readplan.operations.steps-17-to-64"
+	FileReadplanOpsStepsOver64        = "proxyfs.inode.file.readplan.operations.steps-over-64"
 	FileWriteOps                      = "proxyfs.inode.file.write.operations"
 	FileWriteOps4K                    = "proxyfs.inode.file.write.operations.size-up-to-4KB"
 	FileWriteOps8K                    = "proxyfs.inode.file.write.operations.size-4KB-to-8KB"
